add multiple integer sizes

+++ std/io/mod.fs
@@ -1,4 +1,4 @@
-func extern write(fd: i64, buf: *char, count: i64) -> i64
+func extern write(fd: i32, buf: *char, count: isize) -> isize
 
 func print(s: *char) {
     write(1, s, std::string::strlen(s))
